Move base case and remove old cases
