+++ tests/test_drop_tables.py
@@ -10,7 +10,7 @@ import psycopg
 import pytest
 from pytest_cases import parametrize, parametrize_with_cases
 
-from .test_cases.base import Call, MockedResponseTestCase
+from .mock_response_case import Call, MockedResponseTestCase
 
 if TYPE_CHECKING:
     from _typeshed import dbapi
